Revert "Cross platform vite_import plugin (#5513)"
This reverts commit 17f77342f6de39ffbb7733310ea8c75502cfa7ca.
