+++ Dockerfile
@@ -23,9 +23,8 @@ FROM yarn-dependencies AS build
 ADD static/js static/js
 ADD static/sass static/sass
 ADD vite.config.js .
-ADD viteDetectImportPlugin.js .
 ADD tsconfig.json .
-ADD templates templates
+ADD templates .
 RUN yarn install
 RUN yarn run build
