+++ package.json
@@ -19,7 +19,6 @@
   },
   "author": "Canonical webteam",
   "license": "LGPL-3.0-or-later",
-  "type": "module",
   "dependencies": {
     "@babel/core": "7.26.0",
     "@babel/plugin-transform-runtime": "7.25.9",
@@ -33,7 +32,6 @@
     "@dnd-kit/core": "6.3.1",
     "@dnd-kit/sortable": "8.0.0",
     "@dnd-kit/utilities": "3.2.2",
-    "@lvce-editor/ripgrep": "1.6.0",
     "@sentry/browser": "10.12.0",
     "@sentry/react": "7.120.2",
     "@testing-library/dom": "10.4.0",
